sched/walt: wait for cpufreq policies
walt_update_cluster_topology requires the availability
of the cpufreq policies.  These are not available
immediately after insmoding walt, and this causes a
crash through SCHED_BUG_ON.

Register for the cpufreq notifier for POLICY creation,
and init the second half of walt init after the
policy creation for all cpus.

Change-Id: If1af62a4991f3226bd0b661f5a5e419ceea3817a
Signed-off-by: Stephen Dickey <dickey@codeaurora.org>