sched/walt: null ptr dereference fix in walt_lb_active_migration
It is possible that wrq->push_task is null but
walt_lb_active_migration assumes it is non-null. Local variable
push_task is used after copying the value, and can be dereferenced.

Update walt_lb_active_migration to safely check push_task, and
ensure that the active migration flag is cleared.

Change-Id: I3b555430f1ba940053756fcbf2cca5c7d19de9fb
Signed-off-by: Stephen Dickey <dickey@codeaurora.org>