sched/walt: Implement SW cycle counter
Introduce a SW method to count cycles without using hardware registers.

sw cycle counters need to know the current timestamp for which cycles
are account for. It could read the clk register by itself but that is
wasteful, since the walt code already has the wallclock handy. Pass that
along for read cycle counter api. The hw cycle counter variants need not
use the wallclock.

Change-Id: Ie02b027bf1661b3352337adc2fce8e523d0bae78
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Signed-off-by: Stephen Dickey <dickey@codeaurora.org>
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>