sched/pause: support refcounting on pause cpus operation
It is possible that a cpu paused with thermal might be
unpaused by core-ctl, as well as the reverse. This means
that a cpu that has been paused for thermal reasons
can be unpaused when there is need.

Ensure that cpus paused by more than one entity are sticky,
such that only the last unpause request will actually
perform the operation.

Change-Id: Ieb2f836af7348d48a43ae1b22e11d97046ef9dea
Signed-off-by: Stephen Dickey <dickey@codeaurora.org>
Signed-off-by: Sai Harshini Nimmala <snimmala@codeaurora.org>
[WALT ONLY]
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>