sched/walt: Clear scale_frequency_data during init
Change the FIE to use frequency scaling information provided by
cpufreq driver as opposed to the higher priority arch specific counters.

This is done by resetting the struct scale_freq_data for each CPU which
holds the callback information specific to the HW counters, thus forcing
the FIE to revert to default, the cpufreq driver.

Change-Id: I23f508f9859be6a35155e5d000caf1c0b19f5584
Signed-off-by: Sai Harshini Nimmala <snimmala@codeaurora.org>