sched/walt/halt: walt halt functionality
Provide apis to the system in support of halting
cpus, as a replacement for pause/isolation.

Track the requests to halt a cpu, and only
start the cpu again when all requests have been
recanted.

Perform a migration of tasks off of a halted cpu,
if that cpu is online and not idle.

Change-Id: I99340b12e01ff53499b4107d843b7d48e6d6dff1
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>