remoteproc: qcom: Dump md_dbg_buf region on encryption failure
Currently, due to absence of logging framework in modem minidump
encryption driver, no blobs are collected if encryption fails for any
reason. With subsystems adding "md_dbg_buf" region to their public
minidump TOC, add logic in qcom minidump to check for this region as
first blob on encryption failure, and collect it if it is there, instead
of skipping minidump collection altogether. Design is scalable for other
subsystems also.

Change-Id: Ide0f17ca3e7a786f1ee3f7b35fe06705a259e1ee
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Naman Jain <quic_namajain@quicinc.com>
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>