clk: qcom: clk-alpha-pll: Only configure lucid_evo outputs once
Which outputs are enabled when PLL_OUTCTRL is asserted only need to be
configured once. When PLL_OUTCTRL is 0 then all the outputs are disabled
anyway. There's no need to enable/disable the individual outputs each
time the PLL is toggled.

Additionally, in some cases PLLs may be enabled from bootloaders without
all the outputs enabled that we intend to use. In which case we still
need to enable the required outputs during initial configuration even if
the PLL is already on. Enabling additional outputs when the PLL is
already running is safe.

Change-Id: Iffa471783ed5021a92f31673fb2e7539741973b9
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>