sched/walt/core_ctl: create cpu masks for core control
Core Control programatically determins its behavior, in terms
of what other clusters this particular cluster would pay attention
to, to decide whether to unpause a cpu. This is inflexible, and
prevents an easy path to utilizing more clustersto allow unpausing
of cpus to assist clusters in unique ways to support power and
performance goals.

In preparation for using 4 clusters with core control, increase
the flexibility of core control in terms of how it chooses which
clusters to assist, for this cluster.

With this change there are 4 classes of cpus each cluster pays
attention to. Cpus being monitored for nrrun, nrrun misfits,
assist, and assist missfits. This allows the code to replicate
previous core control behavior, while introducing the ability
to specify any cpu to monitor for each of these behaviors.

For example, a typical configuration for a typical 3 cluster
system would be as below, replicating previous core control
behavior but using these masks.

                   nrrun             assist
 cluster | nrrun | misfit | assist | misfit |
 --------------------------------------------
 silver  | 0xFF  | 0x00   | 0x00   | 0x00   |
 --------------------------------------------
 gold    | 0xF8  | 0x07   | 0x00   | 0x00   |
 --------------------------------------------
 prime   | 0x80  | 0x78   | 0x78   | 0x07   |
 --------------------------------------------

This means, for example, when prime computes nrrun it will only
consider itself, for nrrun misfits, golds, and assisting, golds,
and for assisting with misfits, silvers.

This gets expanded to a 4 entry table for a 4 cluster system as
the clusters are fed into core control initialization, allowing
for an extremely flexible definition of how core control operates.

Change-Id: Ic372da9a453ef593144172524ae1d99e056dd45f
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>
Signed-off-by: Abhijeet Dharmapurikar <quic_adharmap@quicinc.com>