net: qrtr: Add dynamic node id configuration
Add support to configure the node id through defconfig. This is useful
for targets that are unable to configure the node id through qrtr-ns
because of security reasons. The local node id can still be overridden
by the ns if it is capable.

Current VM configurations share the defconfig across VMs. This will
cause conflict if multiple VMs are running. Add the ability to set an
override node id from device tree.

Change-Id: Ie9fec2ae276948340f4f5a7e0374d554502a0ee1
Signed-off-by: Chris Lew <quic_clew@quicinc.com>