net: qrtr: Add pm_wakeup_event() support
In some cases system enters into suspend state before handling
incoming packets and cause delay in communication.

Add pm_wakup_event() to abort the suspend when a packet is received.

Force a wakeup when queueing packets to a socket. This is needed to
give userspace clients a chance to run if the device is going into
system suspend. The wakeup event should force the device to abort
suspend.

The qrtr driver can guarantee that the packet gets queued to the socket
but cannot guarantee the client process will get time to run if auto
sleep is enabled. Add a configurable timeout to help mitigate missed
packets on systems where auto sleep is too aggressive.

Use the pm_wakeup_ws_event() api so a hard wakeup can be specified.
This will ensure a resume occurs if the data coming in happens while
the device is going into suspend.

Add support to filter wakeups based on the service ID. This provides
fine control over which incoming packet can wakeup APPS.

Change-Id: Ie69be3a46309f6a05cfbff0642b95194c2d6e955
Signed-off-by: Tony Truong <quic_truong@quicinc.com>
Signed-off-by: Deepak Kumar Singh <deesin@codeaurora.org>
Signed-off-by: Chris Lew <quic_clew@quicinc.com>