DNM: qcom-iommu-util: Code to use max-alignment & best-fit vendor hooks
This patch is a squash of the below patches used to implement best-fit
and max alignment on Pakala.

https://review-android.quicinc.com/c/kernel/qcom/+/5145464
https://review-android.quicinc.com/c/kernel/qcom/+/5145465
https://review-android.quicinc.com/c/kernel/qcom/+/5145466
https://review-android.quicinc.com/c/kernel/qcom/+/5145467

This patch is intented to be used as a hotfix on the current builds
until the ACK changes that are required for this to merge are
available downstream.

The required ACK changes are already merged and should arrive in
the next LTS merge.

Signed-off-by: Oreoluwa Babatunde <quic_obabatun@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>
Change-Id: Ie1aa9bcdc9ebdff8091329c2a1f40cbd532c7b27