soc: qcom: mdt_loader: Handle split bins correctly
It may be that the offset of the first program header lies inside the
mdt's filesize, in this case the loader would incorrectly assume that
the bins were not split. The loading would then continue on to fail for
split bins. This change updates the logic used by the mdt loader to
understand whether the firmware images are split or not. It figures
this out by checking if each programs header's segment lies within the
file or not.

Change-Id: I58686cce70bf1c9d9db8552d31d368d6e61491b4
Link: https://lore.kernel.org/all/20230213185218.166520-7-quic_molvera@quicinc.com/
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>