sched/walt: rearranage busy calculation for cluster
core control should be independently calculating the number of
cpus busy in the cluster, and then using that when evaluating
the need of the system, like every other input into evaluation
of need. Instead, this busy calculation is buried right into
the eval_need code.

Rearrange the code such that the calculation of the busy cpus
in a cluster is contained in a dedicated function, perform the
operation when the rest of the updates are happening, and use
the calculated result in eval_need.

Change-Id: I854c2424108431ed746fb9a3012433b57d98877a
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>