sched/walt: eliminate the nr_assist_thresh
nr_assist_thresh is used by clusters that are assisting other
clusters. Typically this was reserved for gold+, as it's job was to
assist golds. If there were any need for assisting, the cpu would
unhalt, and be available for tasks to run.

With the change to core control to specify a mask for each cluster
that is used to look at which cpus to assist, nr_assist_thresh is
redundant. If the cpus to assist have sufficient tasks, that can
be immediately included in the number of needed cpus.

Update apply_task_need and core control to eliminate the threshold
and always adjust based upon the nr_assist calculated value.

Change-Id: I87f91b67705557c221e2e53fe924e18704ad8be8
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>