i3c-master-msm-geni: Do not change pinctrl state during probe
This change prevents changing pinctrl state to sleep as it creates
problem to enumeration process causing command timeout. During this
phase of driver probe, we expect TLMM to work in IBI mode but changing
to other mode like GPIO or QUP function causes issue. Hence we need to
keep IBI functionality of TLMM during active state and change it to
GPIO mode during sleep state.

With this scenario, AON usecase will be able to control it when camera
switches from HLOS to SSC. This complete change is to satisfy the
requirement of AON camera usecase in mixed bus mode.

Change-Id: I1efca64b1cba1070b4169d426f28a6a1c6446981
Signed-off-by: Mukesh Kumar Savaliya <quic_msavaliy@quicinc.com>