rpmsg: glink: Register for early notifier
Currently when a remote subsystem goes down, remove_subdev will be
called from rproc for removing that subsystem from the list, but
this could be too late. Register for early SSR notifications for
getting notified about before shutdown, so that the glink_send calls
to the dead processor will not get blocked.

Change-Id: Ief7f90caf69f2855272579bc63140837e0efc447
Signed-off-by: Sarannya S <quic_sarannya@quicinc.com>
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>