remoteproc: qcom: pas: Rename subsystem minidump elf name
Currently, the name of ELF remain same whether it is subsystem's
minidump or full dump and it was difficult for parsing tool to
identify the elf's between them.

As previous patch adds the support of creation of ramdump device
with a given name, now we would be creating separate dummy
minidump device for a subsystem with a name appended by _md.
which will help us in creating different name for subsystem's
minidump e.g remoteproc-adsp-md_2022-05-08_18-19-32.elf .

Change-Id: I6b49532b425b480e7351b7d1c86bb0ce29b09e1a
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>