sched/walt/core_ctl: pause rail siblings for single thread work
In the case where the system has a single big task (sbt) executing
the highest frequencies for gold+ cannot be achieved if there
are tasks running on the shared-rail cpus. These cpus need to
go into low power mode for gold+ to achieve single-core-boost,
and this is less likely if the cpus are only partially halted.

The sbt case requires that designated cpus in clusters are fully
halted when sbt is entered and fully unhalted when sbt exits.
This is a distinct client with halt, because the actions are
being taken independently of core control's normal operation.
In other words, core control's decision making process will
have no bearing on whether cpus are being fully halted or
unhalted by sbt, though some sched_avg data will be used,
including the number of big tasks in the system.

Create the sbt check within the core control framework utilizing
existing capabilities like averge task calculations. Enter and
exit sbt mode based upon the number of big tasks in the system
and whether those big tasks are on the prime cpu.

Change-Id: Ibe0bfc72b02cca819551dd9828d2bc4d7f23c70e
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>