soc: qcom: Add PMIC Glink driver
The PMIC Glink driver provides the interface for clients to
communicate over GLink for sending and receiving data to charger
firmware that runs on a remote subsystem (e.g. ADSP) which
supports charging and gauging.

This enables clients to read/write battery charging parameters
and get notifications.

This is a snapshot taken as of qcom-6.1
commit b57cad08fbf8 ("soc: qcom: pmic_glink: use semaphore
in pmic_glink_register_client()").

Rename pmic_glink.c to qti_pmic_glink.c and pmic_glink.h to
qti_pmic_glink.h to avoid a conflict with the independent
upstream implementation of these files.

Change-Id: I1e4c09d1ee10c8c5397f4459afe8d656593d718a
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>