usb: typec: ucsi: add QTI UCSI Glink driver
Qualcomm Technologies, Inc. UCSI Glink driver handles the
communication between OPM on the Application processor and PPM
which is charger firmware running on the remote subsystem
(e.g. DSP) over PMIC Glink.

This is a snapshot taken as of qcom-6.1
commit 61257d5c141f ("usb: typec: ucsi_glink: Add charger
remoteproc timestamps in logs").

Rename ucsi_glink.c to ucsi_qti_glink.c to avoid a conflict
with the independent upstream implementation of this file.

Change-Id: If7fced3eebee9d41ad1f51229c28e4f63b750250
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>