regmap: add QTI regmap debugfs library
The current method of reading register files dumps the entire
address space.  One can use the 'dd' command to dump a subrange
within the address space.  However, that requires knowledge of
the string length of each line which is derived from the max
address, the character length of each register entry, and the
format.

Provide a simple means to dump a range of registers by allowing
the user to specify the start address and the count of registers.
When the data is read convert the dump address to a starting
position in the file.  Similarly if the file offset goes beyond
the dump range return 0 to indicate that the data is already
dumped.

Also provide a means to write to a register address.  Expose
register and unregister functions in a library that can be
invoked by drivers to take advantage of this fine-grained
debugfs register read/write interface.

This is a snapshot taken as of qcom-6.1
commit f056d70d999f ("regmap: qti-regmap-debugfs: correct
register reading for large count").

Change-Id: I7004c0378225c13e8e31c21b961f6e72814f2a58
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>