spmi: add SPMI PMIC arbiter debug bus driver
Add a new SPMI bus driver which can manage the SPMI debug bus
found on SPMI PMIC arbiter version 5 and above.  This debug
bus has read and write access to all PMIC peripherals regardless
of ownership configurations.  It cannot be used on production
devices because it is disabled by an eFuse.

This is a snapshot taken as of qcom-6.1
commit 2b37abe39681 ("spmi: spmi-pmic-arb-debug: add support
for fuse enable bit").

Change-Id: I7dd7171d775db02269b98bb079bdccf5d0f4aae4
Signed-off-by: David Collins <collinsd@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>