sched/walt: create an api for cluster boost for pipeline
Not all clusters should be unhalted when pipeline is enabled,
but all clusters are being unhalted now (through
core_ctl_sched_boost).

To prepare for only unhalting the clusters that are part
of the user selected pipeline cpus, create a per-cluster
boost api which controls the boost characteristics of
only the cluster idx specified.

Change-Id: I92fdc13bcb6ee84782b5ea1b5dcb74caabee8ccf
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>