soc: qcom: add PMIC PON log parser driver
Add a driver which can parse PMIC PON data into a human readable
format.  This log is stored in a binary data structure in PMIC
NVMEM starting with PMIC PMK8350.  It contains PMIC power-on,
power-off, fault, and other information through multiple power-
cycle events.

This is a snapshot of the pmic-pon-log driver taken as of qcom-6.1
commit cc0b5d105101 ("soc: qcom: pmic-pon-log: update pon trigger
map for USB").

Change-Id: I75a7239f8927b453676a89db6e28c2ba413c30bd
Signed-off-by: David Collins <quic_collinsd@quicinc.com>