soc: qcom: add qti_battery_debug driver
The QTI battery debug driver helps to get logs and debug information by
communicating with charger firmware (CHGFW) running on the remote
subsystem (e.g. DSP) over PMIC Glink. It supports getting the device
context dump of battery gauging software which is part of CHGFW.

This is a snapshot of the qti_battery_debug driver taken as of qcom-6.1
commit ad98e248e2cb ("soc: qcom: qti_battery_debug: add support for 2S
battery context dump").

Change-Id: I37460f7a17038f1798a221c96a87649cb65dd357
Signed-off-by: David Collins <quic_collinsd@quicinc.com>