soc: qcom: add PMIC glink debug driver
Add a PMIC glink debug driver for the Glink PMIC register dumping
debug interface.  This can be used to read and write PMIC registers
controlled by the charging firmware running on a remote processor
(e.g. DSP).  The PMICs may be physically connected using either SPMI
or I2C.

This is a snapshot of the pmic-glink-debug driver taken as of qcom-6.1
commit 93d52d6c433e ("soc: qcom: pmic-glink-debug: Fix initialization
in i2c_glink_xfer()").

Change-Id: I159afd997abb0c4f4889c34e39e026dede06ceb4
Signed-off-by: David Collins <quic_collinsd@quicinc.com>