scsi: ufs: ufs-qcom: Remove conflicting code
As a result of the upstream change that makes scsi_host_template asdf
pointer a constant, write access to this structure is prohibited in the
UFS qcom driver. Remove the re-initialization of the
rpm_autosuspend_delay field of the mentioned structure in ufs qcom
driver. In the current UFS PM implementation, the host is the parent of
the UFS device's LUNs. The parent cannot go into suspend unless all the
children are in suspend. Therefore this host's rpm_autosuspend_delay
re-initialization in the qcom driver is unnecessary.

You can refer to the corresponding upstream change below:
https://git.kernel.org/pub/scm/linux/kernel/git/stable/linux.git/commit/?id=3b465a1491

Change-Id: Id1818c308912addc6d7695da6b111bbfebddb18d
Signed-off-by: Bao D. Nguyen <quic_nguyenb@quicinc.com
Signed-off-by: Blagovest Kolenichev <quic_c_bkolen@quicinc.com>