sched/walt: drain_rq_cpu_stop should unpin the rq lock
When drain_rq_cpu_stop calls __balance_callback, a subsequent
operation for pushing a task, can unlock the runqueue.

 lock_release
 _raw_spin_unlock
 raw_spin_rq_unlock
 find_lowest_rq
 push_rt_task
 push_rt_tasks
 __balance_callbacks
 drain_rq_cpu_stop

This is causing a "releasing a pinned lock" warning, because
drain_rq_cpu_stop locks and pins the rq lock.

Address this issue by ensuring that the rq lock is unpinned
upon the call to __balance_callbacks, and that before exiting
the pin state of the lock is no longer checked.

Change-Id: Ib0ce4a2954567631b49c52c94bb6a0ce25e3153f
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>