regulator: add QTI regulator OCP notifier driver
Add a qti-ocp-notifier driver which is used to log and notify
consumers of regulator over-current (OCP) and alarm events.
This driver is needed to monitor regulators controlled by RPMh
since the application processor is unable to directly receive
OCP interrupts for RPMh managed regulators.

A regulator alarm event occurs when the output voltage droops
due to excessive current consumption which is lower than the
over-current protection threshold.

This is a snapshot of the qti-ocp-notifier driver taken as of
qcom-6.1 branch
commit af3266829e1a ("regulator: qti-ocp-notifier: rate limit
OCP and alarm notification logging").

Change-Id: I1fac154a5c6f70454ed1a013b9013c208d02dcce
Signed-off-by: David Collins <quic_collinsd@quicinc.com>