remoteproc: qcom: q6v5: Do not stop glink subdevices on recovery disable
When recovery is disabled, all the subdevices of remoteproc gets
stopped so that they can collect data in their ssr notifiers
but it also stops glink subdevice at the end, which cleans up
the data which is useful to debug issue related to glink
communication with the remote.

Let's put a check in glink stop about remoteproc recovery status
and if it is a crash and not a graceful shutdown.

Change-Id: Ib8314d0d8aa0c4565e0018bba093d3b0f2009cdd
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>