input: misc: add QTI HV haptics driver
Qualcomm Technologies, Inc. (QTI) HV (High Voltage) haptics is a
PMIC module which can play haptics effects for LRA or ERM with drive
voltage up to 10 V. It has several pattern sources built inside and
each pattern source can be used for playing different vibration
effects.

Add the driver to support QTI HV haptics module, which will be
registered as an input FF device and can be controlled by calling
read/write/ ioctl APIs on the input device node.

This is a snapshot of the qcom-hv-haptics driver taken as of qcom-6.1
commit ce16d31c8217 ("Merge "input: misc: qcom-hv-haptics: adjust
VMAX headroom according to VMAX"") with debugfs and trace features
removed.

Change-Id: I99241638eac57c5e02d2bcb1d0e27bd320818ede
Signed-off-by: David Collins <quic_collinsd@quicinc.com>