mfd: introduce I2C PMIC controller
The I2C PMIC Controller is used by multi-function PMIC devices
which communicate over the I2C bus.  The controller enumerates
all child nodes as platform devices, and instantiates a regmap
interface for them to communicate over the I2C bus.

The controller also controls interrupts for all of the children
platform devices.  The controller handles the summary interrupt
by deciphering which peripheral triggered the interrupt, and
which of the peripheral interrupts were triggered.  Finally, it
calls the interrupt handlers for each of the virtual interrupts
that were registered.

This is a snapshot of the qcom-i2c-pmic driver taken as of qcom-6.1
commit 86c6bfd9ab12 ("mfd: qcom-i2c-pmic: Change the return type
for remove callback") with debugfs features removed and pr_*()
log statements changed to dev_*().

Change-Id: I9aae0d7ae25a76e4a644c8aebc0099325cbac1e8
Signed-off-by: Nicholas Troast <ntroast@codeaurora.org>
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>