hwmon: Add QTI AMOLED ECM driver
AMOLED ECM (Embedded Current Measurement) is a method to measure OLED
display power consumption with a time granularity in sub-frame or
multiple frames of image data. ECM can be for a short period or for a
longer period.

This is a snapshot of the qti_amoled_ecm driver taken as of qcom-6.1
commit 131656a2b76c ("hwmon: qti_amoled_ecm: Support getting
sdam_index for more than 2 SDAMs") with pr_*() log statements changed
to dev_*().

Change-Id: I037d490e170cde75445a083950fe7ba7628c8078
Signed-off-by: Shyam Kumar Thella <sthella@codeaurora.org>
Signed-off-by: David Collins <quic_collinsd@quicinc.com>