firmware: qcom_scm: Modify WAITQ_WAKE handling
Currently when handling a WAITQ_WAKE, the wake_ack being sent as part of
the handshake sequence may itself receive a WAITQ_WAKE. If this second
WAITQ_WAKE contains a _different_ wq_ctx from the one received
previously for the first, we will never call flag_handler() on the first
wq_ctx. Thus, we will miss out on waking up an SCM call sleeping on the
first wq_ctx, which is bad.

To fix this, move the flag_handler() to get called as soon as a
WAITQ_WAKE is received, right before the wake_ack is sent.

One may well wonder whether this subversion of order is okay - it is;
the underlying firmware doesn't care about whether a sleeping request is
woken up via wq_resume() before or after the wake_ack().

Change-Id: Ifbd87bf696c6e56004fa4305514247e36fa60f12
Signed-off-by: Guru Das Srinagesh <quic_gurus@quicinc.com>