input: misc: qcom-hv-haptics: add MMAP support for HAP530_HV
HAP530_HV haptics module increases the total FIFO memory space to 8192
bytes for both FIFO and PAT_MEMx mode usage. It also updates the
PAT_MEMx partition management with a flexible strategy such that just
the start address and the length only needs to be programmed into
corresponding registers before updating and playing the waveforms
inside a partition. The number of the partitions is also flexible (no
longer fixed to 4) and there is no limitation to each partition size
as long as there is enough memory space for new partition allocation.
However, the 1st 640 bytes of memory space should be always reserved
for FIFO partition, and any memory space that hasn't been allocated
between the FIFO partition (from the beginning) and PATx_MEM partitions
(from the end) will be finally configured as part of the FIFO partition
for better sample refilling efficiency during FIFO mode play.

Change-Id: Icf008de3c4addad00e392b01f9cff5c85504f31c
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>