abi_gki_aarch64_qcom: Update symbol table to add wait_for_device_probe
The DWC3 MSM driver requires a strict parent and child probe sequence, and
this can lead to issues if probe is done out of order.  In scenarios where
the DWC3 child takes the async probe path, the parent is unable to
determine if the child probe routine completed with a failure, or simply
deferred.  wait_for_device_probe will allow the parent to wait for the
device's probe to complete before continuing to check if the probe was
successful.

Change-Id: Id22e225a365c37b3fd804c88578d4964856e2de5
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>