modules.list.msm.sun: Move usb drivers to second stage
If we keep the usb drivers in the first stage, it gets probed ahead of
iommu driver and that leads to crash. If we move this to second stage
the module loading will be delayed and the usb driver load will come
after the ufs drivers are probed (second stage modules are coming in
from /vendor_dlkm partition) ufs driver has a dependency on iommu driver
and thus we can hide this issue. We need to go back and check why the
module probing is not correctly ordered.

Change-Id: I7d8d238a60e7b1687d4e12f43146310d9dbb28f8
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>