pci: msm: Fix the drv_supported flag use in pcie driver
Fix the way drv_supported flag is used by tying it with the
qcom,drv-name dt flags instead of supporting drv by default
when pcie cesta is available.

With these new changes, drv_supported flag will be true only if
qcom,drv-name = "lpass" or "cesta" dt flag is present.

Fix couple of debug logs format to align with the rest of the logs.

Change-Id: I36f366f126ce663ba76eff356bde61dedf0b33a9
Signed-off-by: Prudhvi Yarlagadda <quic_pyarlaga@quicinc.com>