sched/walt: update to fix general coding style errors
In general when a patch is submitted, checking tools only
perform tests on the changes being applied, rather than an
entire file. This has allowed checkpatch errors and warnings
to creep into the walt source code.

Note that none of these changes will impact code execution, excepting
the replacement of strlcpy with strscpy.

Change-Id: I19626b2e135d6c034fc381b635eba39bdbc92b40
Signed-off-by: Stephen Dickey <quic_dickey@quicinc.com>
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>