firmware: qcom_scm: Allow DMA to be allocated anywhere in DDR
For VM bootup the scm device needs to have memory from above 4GB, but for
PIL boot the Metadata memory needs to be limited to 32 bit. This is
because the authentication software for the metadata in the secure world
works with only 32 bit addresses. Add support for only 32 bit addresses
for PIL and memory from anywhere for other memory allocation from SCM
device.

Currently, this is being enabled for ADSP/CDSP/MSS however, this limitation
 was applicable only for modem and not with ADSP/CDSP but there was some
issue observed with above 4G addresses allocated for ADSP/CDSP and it
was analyzed that it could be only issue on emulation platform and will
not observed on Silicon. So, we could revert this change for ADSP/CDSP
if the issue is not observed on Silicon.

Change-Id: I4dfc6bebfabdcdd31e72903525f5f298c918c416
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>