ANDROID: build: Fix system_dlkm image input to super.img
Currently, we use "system_dlkm.img" as an input to "super.img".
With latest LTS, the kernel_images rule is able to produce multiple
system_dlkm images based on whichever file system types are defined.
This breaks the super image rule because it can only take one file
as an input for system_dlkm.

Update such that we use system_dlkm.ext4.img for perf builds, and
system_dlkm.img for consolidate builds.

Change-Id: I02b8434de10cf321aa6edab4d2c35e5480482d9d
Signed-off-by: John Moon <quic_johmoo@quicinc.com>
Signed-off-by: Blagovest Kolenichev <quic_c_bkolen@quicinc.com>