sched/walt: Fix condition for bigger number of clusters
In a system of 2 clusters or less, as num_sched_clusters is an unsigned
int, doing a x-3 operations is going to result in a wrapped around
large number. Furthermore, this condition was only specific to one
platform. Fix this by taking advantage of WALT config to ensure the
end_index is only set for the appropriate target.

Change-Id: Ibb912941aa93508e1cbd2825ac5f479c19728c45
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>