hwmon: qti_amoled_ecm: add support for PMD802X
PMD802X supports AMOLED ECM which is also known as average current
detection (ACD). On PMD802X, this feature doesn't need to be
enabled or disabled at runtime via AB peripheral (ELVDD in PMD802X)
as in PM8350B. This is always enabled and handled by PBS. Also,
the gain factor has been increased to 50 mA. Add support for it.

Change-Id: I10e8c71a4bc53fedcd96de83fc4a571ab4eb53b0
Signed-off-by: Subbaraman Narayanamurthy <quic_subbaram@quicinc.com>