remoteproc: soccp: Indroducing SoCCP
The SoC Control Processor (SoCCPP) is small RISCV subsystem providing
support for general SoC infrastructure management that cannot be
effectively managed by the CPU or larger DSPs. It provides a solution
for control-plane processing, reducing per-subsystem microcontroller
reinvention.
SoCCP driver borrows the unprepare/start/stop/load/coredump callback
from the pas client driver and in addition adds an API (rproc_set_state),
this API is exposed to the SoCCP client drivers to control the power
modes of SoCCP. The power modes controlled by the clients are dormant
and active.
rproc_set_state is refcounted and the request to set to dormant state
is sent to SoCCP only if there are no active clients for SoCCP.
set_state_soccp is a blocking API which polls on the config register for
the requested state change.

Change-Id: I91347b34700c76efe8d4e3086062fa59ba2f9583
Signed-off-by: Gokul krishna Krishnakumar <quic_gokukris@quicinc.com>