input: misc: qcom-hv-haptics: Update Vmax setting for impedance detection
Vmax needs to be set to 10V in binary sweeping impedance detection and
it was done by listing VMAX register as part of a group of register
settings with corresponding raw data values. This needs to be updated as
the Vmax step size is changed in different variants, so add a helper
function for Vmax programming and call the helper function to have it
handle the different Vmax step cases.

Change-Id: I07dc44e9d8eb2fea83e7895f25cd0c2860d4a8ab
Signed-off-by: Fenglin Wu <quic_fenglinw@quicinc.com>