usb: dwc3: dwc3-msm-core: Update EBC USB request TRB parameters
When queueing EBC requests, the DWC3 MSM overrides the default endpoint
queue callback and implements sequences required to start the transfer.
The callback is expected to maintain the DWC3 TRB parameters, such as
num_trbs and trb_enqueue, so that standard DWC3 gadget calls can update
them accordingly.  In the dequeue path, the design utilizes the exisiting
dwc3_gadget_ep_dequeue() API, and expects that these are properly
populated.

Fix up the current implementation to increment trb_enqueue and
num_trbs, so that the cleanup path resets the endpoint, and a new session
can be started.

Change-Id: I3bb1b96a6aee0de70e340b04288083d1171fc2b8
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>