interconnect: qcom: sun: Mark SN4 as keepalive_early
This node is only used for PCIe and it's voted for by default coming out
of bootloaders. If PCIe never votes for it through interconnect, then
we'll just leave the bootloader vote in place forever and block system
low power modes. Mark it as keepalive_early, which will cause us to
remove the vote in sync_state, but won't always keep it enabled when
HLOS is up.

Change-Id: Iaef1dd949885601d86f646711904f2a9e2619dbc
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>