arm-smmu: Support per-device fault handlers
The iommu framework allows registering a device-specific fault handler
via iommu_register_device_fault_handler(). When a fault occurs on a sid
matching this device, call the handler.

This could be useful if multiple devices share the same iommu group.

The old fault handler mechanism is left as is, and clients choose to
use either.

Change-Id: I7905f35372805886a4962c620a3e9f6403f07bfe
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>