firmware: qcom_scm: Move shmbridge init from module_init to scm probe
Calling shmbridge initialization from scm driver module init
does not always guarantees that scm driver would have been
ready at the time of starting shmbridge initialization as
there is a chance scm is waiting on resource and it might
not be probe at the time of module init and that can make
NULL pointer dereference issue for scm bridge initialisation
and during its probe.

Shmbridge driver is only depended on scm driver readiness
so that it could make an scm call, for that reason move
its init call to scm driver probe.

Change-Id: I25307d9652a5627d3dd46e6a23a32c7230b8784e
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>