firmware: qcom_scm: Use Scm device for scm calls
Commit 6b3b78c53b88 ("firmware: qcom_scm: Allow DMA to be
 allocated anywhere in DDR") replaces scm device which
should be used for scm call by a passed 32bit device with
purpose to allocate dma memory in lower 4GB range but
scm device should be used for all the necessary data
needed to make an SCM call.

Fix this by using passed 32bit device only for dma memory
and not for scm call.

Fixes: 6b3b78c53b88 ("firmware: qcom_scm: Allow DMA to be allocated anywhere in DDR")
Change-Id: Ib2f0b9f293c0109967d61cea5822f2a84dffbbb6
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>