interconnect: qcom: bcm-voter: Use new crm_write_bw_pt_vote() API
CESTA 1.0 didn't support CRM instances with both a CRMB and CRMB_PT.
They could have one or the other, but not both at the same time.  Thus,
the CRM driver originally supported both of them through the same
crm_write_bw_vote() API. However, CESTA 2.0 added support for instances
with both a CRMB and CRMB_PT. And due to this, one needs to distinguish
whether the node indices are for the CRMB or for the CRMB_PT.

To support this, the CRM driver added a new API dedicated to the
CRMB_PT. The old API will only be used for the CRMB. Since interconnect
currently only supports paths using the CRMB_PT, switch to the new API
unconditionally.

Change-Id: I9388fcd4c3befa99733740a69af1fb65feab6169
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>