soc: qcom: memshare: Restore scm changes
Parts of the memory sharing operation were left out in the initial
snapshot because dependencies weren't ready. This change restores the
functionality to call hyp_assign and register regions with rpmsg glink
for the zero copy feature.

With this change the only missing functionality is remoteproc trace
events that are used to debug ssr event timings.

Change-Id: I7636583488e89d3292cda8694467bf6b9fa5b46e
Signed-off-by: Chris Lew <quic_clew@quicinc.com>