remoteproc: sysmon: Fix the broken sysmon ssr state machine
Currently the state machine of sysmon ssr is in broken state
due to two variables sysmon->ssr_event and sysmon->state
being used for the same reason and they are not properly
set at appropriate place and due to which one subsystem is
not aware about other subsystem state(UP/Down) that can
result in unexpected crash of other subsystem system
like e.g ADSP crashes for some reason and its status is
not known to MSS and due to which MSS as well going to crash.

There is one more bug where ->state_lock is not being
released in sysmon_start() if ssr subsystem is successfully
able to send sysmon event to the remote.

Let's fix both the issue with this change.

Change-Id: Ib2b311496d4f2f06f2089bc50b49374c4779e22b
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>