clk: qcom: reset: Ensure resets trigger before returning
The reset control signals are sent through demets, which introduce
delays between writing the control registers and the reset operations
actually triggering. We aren't waiting for the demets before returning
to the client, so the resets may not be triggered yet in all cases. The
client should be able to assume the operation has actually completed
after calling reset_assert() or reset_deassert(). Clients should only be
responsible for ensuring the necessary delay between the assertion and
de-assertion -- not any internal clock controller delays for triggering
the resets.

Additionally, writes to these registers may be posted, which means the
bus may issue an early write ACK before the write is fully propagated to
the control registers. We aren't reading the register again after
performing the write, so we might return to the client before the write
has actually completed.

Ensure the resets are actually triggered before returning to the client
by reading back the register after writing, and adding the necessary
demet delays.

Change-Id: I77b4632f3250d655e663bbc405736965ef2d6bad
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>
Signed-off-by: Vivek Aknurwar <quic_viveka@quicinc.com>