sched/walt: Limit sysctl output to number of clusters
The system can accommodate chips with up to 4 clusters. For sysctl nodes
such as sched_fmax_cap, up to 4 entries of information can be displayed.
However, if an SoC has fewer than the maximum number of supported
clusters, it is unnecessary to display this excess information.
Therefore, ensure that each sysctl node only displays the information
that accurately conveys the message for the number of clusters
associated with that SoC.

It is important to note that when writing to proc_do*vec, the buffer
for the ctl_table must always match in size to the input buffer as
defined in the sysctl interface. In the event that the temporary buffer
size is less than the input buffer, unexpected behavior can be observed,
namely in nodes being reset to 0.

Consider the example of sched_fmax_cluster. The temporary buffer size is
reflected by cap_margin_levels (set to num_sched_clusters in this
context), which would be 4 in the case of a 4 cluster system. Note
the size of the input buffer in the walt_table for this sysctl node
is defined by MAX_CLUSTERS which is also 4. However, in a 2 cluster
system, cap_margin_levels would now reflect 2, while the input buffer
size remains constant at 4 (to maintain backwards compatibility).

This creates a situation where the temporary buffer size is smaller
than the input buffer size. In this case, the following behavior was
observed:

dmesg logs after doing:
echo 1 2 3 4 > sched_fmax_cap

print: entering handler
print: writing 1 2
print: sched_fmax_cap: 1 2 unset unset
print: ret = 0

print: entering handler
print: writing 0 0
print: sched_fmax_cap: 0 0 unset unset
print: ret = 0

The handler was being called twice, and the end result was
sched_fmax_cap having its nodes reset to 0. To avoid this,
ensure the input buffer size and temporary ctl_table writing buffer
size are always equal.

Change-Id: I4b556e653b04c173b765ff659da40e6b442c74aa
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>