clk: qcom: gcc-sun: Don't overwrite CESTA PCIe rates in sync-state
If the PCIe clocks are enabled from bootloaders, then sync-state will
disable them in sync-state if nobody votes for them through the clock
framework. However, at this point CESTA may already be controlling the
PCIe clocks. We want sync-state to remove any HLOS-specific votes for
the PCIe clock branches, but we don't want it to overwrite the RCG rates
due to safe-config parking. So, disable the safe config parking for the
PCIe RCGs that are CESTA-controlled to prevent this.

And disabling parking means clk_set_rate() won't short-circuit when the
RCG is disabled anymore. Which means we don't guarantee the parents
being switched to are on at the time, either. If HW_CLK_CTRL is set,
this will cause the RCG to get stuck in UPDATE, since that requires a
dynamic, glitch-free switch. To prevent this, clear HW_CLK_CTRL to allow
static switches that don't require the parents.

Disabling the parking behavior is safe for this RCG, since nothing other
than SW (or CESTA when it's controlling it) can enable any clocks
downstream from it. There are no HW signals that can enable its clocks
while the RCG is configured to a disabled parent.

Change-Id: If6be6abfbdfb6db53c5cb19612586f01f0f1ab81
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>