clk: qcom: clk-alpha-pll: Don't skip writing values of zero
Zero is a valid register setting in some cases. If HW defaults the
register to a non-zero value, then we must explicitly write it to the
expected zero setting.

The current code skips zero-value writes so the pll_configure()
functions can be shared across PLL generations with different numbers of
configuration registers. This works because if the register isn't
supported, then their values should be zero in the alpha_pll_config
structs. But this doesn't work the case above when zero is actually the
expected value.

Instead of relying on a zero value to check if the register exists, use
the register offset from clk_alpha_pll::regs instead. The only register
with a zero offset is the MODE register, which isn't used for these
kinds of settings. Any valid configuration register will have a non-zero
offset.

Change-Id: I604f90322f3cd2d0cf8f9a5601587c0410819f92
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>