sched/walt: Allow active load balance of halted similar cap CPU
Consider the case of a prime cluster with 2 CPUs, and both are halted. A
task which is affined to both CPUs starts running. At window rollover,
one CPU, CPU X is unhalted, while the other (CPU Y) is still halted.

Consider further that the task had woken up on CPU Y.

Since CPU X is unhalted, the task should be migrated from CPU Y and made
to run on CPU X, as the expected behavior is for tasks to not run on
halted CPUs if a worthy unhalted CPU is available. This is supposed to
be handled via the walt_newidle_balance path.

However, there are two roadblocks to this in the code presently.The
first is recognizing a busy CPU. When searching for busy CPU candidates in
the same cluster, only CPUs which are running more than 2 CFS tasks are
considered. In the case described above, however, there is only 1 CFS
task running on CPU Y.

To fix this, create an exception when searching for a viable busy CPU
candidate in the same cluster as the newly idle CPU. Namely, ensure that
if a CPU is currently halted and has a task running on it, it is at
least considered as a potential candidate for busy CPU (for now, allow
for other CPUs to be considered busy CPU candidates in the same cluster
if they are more busy).

The second roadblock is in the active load balance path, once a busy CPU
has been selected (which is now referred to as the source CPU, while the
newly idle CPU is referred to as the destination CPU). If the source CPU
is currently running a task, it will only be pulled onto the destination
CPU if the destination CPU has a higher capacity than the source CPU. In
the case described above, this will never happen, as the prime cluster
is the highest capacity cluster.

To fix this, allow active load balance to pull a task from the source
CPU to the destination CPU in the event that both the source CPU and the
destination CPU belong to the same cluster, and that the source CPU is
halted.

Change-Id: I6ce171a651a152080a1d8db7a1ac1dbe436befea
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>