sched/walt: Ensure proper update of nr_need
compute_cluster_nr_run is called on a cluster, and is should return the
number of tasks running across that cluster, as well as any other
monitored clusters, as defined by userspace via the nrrun_cpu_mask.

In an earlier target, commit e35d899610ee
("sched/walt: prevent early unhalt of titaniums") was
introduced as a "fix" for an issue observed due to partial halt CPUs.
If a cluster A contained CPUs that were partially halted, and that
cluster was also monitoring the prime cluster (which happens to be running
a task), then due to task running on the prime cluster, the nr_need
associated with the cluster A would have a need of 1. However, this
would lead to the partially halted CPU on cluster A coming out of
partial halt and being fully unhalted, even though there was no need
of this.

While that change works fine for clusters that host partially halted
CPUs, it breaks when partially halted cpus are not involved, and
specifically when the highest order cluster is hosting more than one
CPU. Consider the example of a topology which has two primes cpus, where
one cpu is unhalted and busy with a task, while the other CPU is in
halted state. When compute_cluster_nr_run is called, the cluster's need
will be recognized as 1, due to the busy task, however, immediately be
decremented by the number of active CPUs in the mask, and so the number
of unhalted CPUs would be the one unhalted prime. Therefore,
compute_cluster_nr_run would return 0.

This can be problematic, because until the task is busy enough to
trigger nr_busy being incremented to 1, the overall nr_need will read 0,
which could lead to a request to halt the active prime CPU, despite the
a task actively running on it.

Note that nr_busy takes in to account the freq invariance while nr_need
is simply a reflection of the busy time on that cpu. In other words,
nr_need could be 1 even if the cpu is running at its lowest frequency
while nr_busy will be 0, because it is utilization based.

Fix this by ensuring compute_cluster_nr_run only decrements the active
cpu count if and only if the cluster on which its called is currently
hosting partially halted CPUs.

Change-Id: I1d0225b7acf59854d2eedd5f5f2bb2a3f48099d1
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>