sched/walt: Thermals based oscillation on Prime CPUs
Architectures that have multiple big cpus could reduce
thermal residencies for single big task usecases by rotating the tasks
amongst them.
Detect if there is a One Big Enqueue Task type of load on Prime CPUs and
continue to rotate the single big thread between Prime CPUs in a
periodic fashion if thermals are occurring on largest cluster.
Detection to enter into oscillation is carried out every tick and then
once oscillation starts, exit conditions are checked right before the
task oscillates on to the next CPU in addition to every tick. If
oscillation stops, it will not be re-enabled until the next tick where
the entry checks are made.
Ensure that LPM bias on largest cluster is disabled when OBET mode is
detected, to benefit from single core cpufreq boost.
Ensure that core ctl boost is set for largest cluster when OBET mode is
detected to prevent isolation of cores during oscillation.

Change-Id: I48540e64097a19ddb9c02a627cf10ccc359515d3
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>