drivers: soc: qcom: Use key retention in ICE
ICE has a key retention feature that keeps the keys in their slots
after a pause/resume. If we do not have to reprogram the keys, we should
not rewrite they key into its slot as this increases latency. This
change checks the registers to see if the key is still enabled, and if
so, immediately returns during program_wrapped_key.

Change-Id: I37db3494ab7c6b6f1ff027a461ba7f49b6755932
Signed-off-by: Daniel Perez-Zoghbi <quic_dperezzo@quicinc.com>