clk: qcom: gdsc: Add support for HW_CTRL_SKIP_DIS flag
The current gdsc_disable() callback switches the GDSC back
to SW mode and then disables the parent supply. But display
driver requires the GDSC to be always in HW control mode,
but the vote on GDSC's parent supply needs to be removed in
gdsc_disable() callback. Introduce HW_CTRL_SKIP_DIS flag which
can be added for display GDSC's to achieve above functionality.

While at it, fix gdsc_disable() logic to remove vote on regulator
parent supply for votable GDSC's.

Change-Id: I41c49628ab694a9dfb200eb159c4a12c297ac892
Signed-off-by: Jagadeesh Kona <quic_jkona@quicinc.com>