sched/walt: Introduce trailblazer state
Introduce a method to track how long a trailblazer task has been
running in a window. The method to do this is very similar to how
misfits are tracked.

First, there is a running per CPU counter which tracks the number of
trailblazer tasks that are on that CPU's runqueue. Next, a prod sum is
calculated to determine if the trailblazer task is running for at least
60% of the window.

Whenever there is a CPU where a trailblazer task is running for at
least 60% of the window, the system will be placed in a
"trailblazer state". Under trailblazer state, the floor of the gold
frequencies will be raised to 1Ghz. This setting will be configurable
per SOC, and can be expanded further to allow different floors for
different clusters.

Change-Id: Ieffebd56ea4d5b1b2961c4ee6c90ca23fd824ee7
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>