stm: class: fix slab-out-of-bounds in ost_t_policy_entity_store
The p_ost driver recently added a private member ost_t_policy_node,
if the priv_sz of is not set, when the p_ost policy is created, no
memory is allocated for this private member. then cause this issue.
This change sets priv_sz to sizeof(struct ost_t_policy_node) to
fix this issue.

Change-Id: I4506718da544a357f54e12072b9e1b1e80878cfc
Signed-off-by: Jerome Lee <quic_jaewookl@quicinc.com>
Signed-off-by: Yuanfang Zhang <quic_yuanfang@quicinc.com>