soc: qcom: mem-hooks: Register hook for android_rvh_try_fixup_sea
path_lookupat() is capable of safely reading unmapped VAs, by using
load_unaligned_zeropad(). If an unmapped VA is read whilst the
function is being called, that resulting page fault will get
re-directed to __do_page_fault(), which will call fixup_exception() to
handle the aforementioned unmapped-VA-read.

Now consider some memory was still mapped at S1 but lent to another
Gunyah VM. path_lookupat() can then use load_unaligned_zeropad() on a
page that virtually borders - in the S1 VA space - the page that was
lent to the other VM, and end up trying to read content from that
page. Gunyah will then inject an syncrhonous external abort for a PTW
fault, which will map to do_sea(). Register a callback for
trace_android_rvh_try_fixup_sea(), which returns true if we can
attempt a fixup_exception() call in do_sea() to zero-pad the read to
the page that was lent away.

Change-Id: I30a85d33f6c52d902275a178191cd47a2336658a
Signed-off-by: Chris Goldsworthy <quic_cgoldswo@quicinc.com>