ufs: ufs-qcom: Change the PHY init sequence to power on first
In the current initialization sequence, the UFS PHY is reset
before the power is turned on. Resetting the PHY without its power
may cause incorrect behavior from the PHY.

In the updated UFS initialization, first turn on the PHY power before
issuing the UFS host controller and PHY reset. If the UFS PHY
power has already been turned on earlier, turning the power on
again would result in a no-operation.

Change-Id: I96c4456358aecd1ca1c197e4cfba0b04c1dc7aba
Signed-off-by: Bao D. Nguyen <quic_nguyenb@quicinc.com>