sound: usb: qmi: Remove XHCI endpoint during ADSP SSR situations
In the path of when the QMI handle is removed, i.e. due to ADSP SSR, the
stream disable request is not received in the USB SND QMI driver.  This
would result in the XHCI endpoints not being removed, leading to failures
when requesting/adding the endpoint on subsequent session starts.  Ensure
that XHCI endpoints are removed within the path of when the QMI handle is
removed.

In addition, remove the need to run the QMI disconnect routine as a
separate workqueue, as the QMI interface already runs the QMI client/server
removal sequences in a workqueue.

Change-Id: Iafe8f9295b07278c9a078843304e80cb12f9d850
Signed-off-by: Wesley Cheng <quic_wcheng@quicinc.com>