dma-mapping-fast: Fastmap input buffers must meet the alignment requirements
The kmalloc() minimum alignment on arm64 has been reduced from 128 to 8
and some additional bounce buffer handling have been introduced with
commit 861370f49ce4 ("iommu/dma: force bouncing if the size is not
cacheline-aligned")

Make sure that fastmap input buffers are meeting the alignment requirements
or otherwise fail the mapping and print a warning.

If we want to add bounce buffer support to fastmap, then the changes from
the above commit should be replicated in the fastmap code.

Change-Id: I4e6c1803f63d238591b68d88787444553a9bf90c
Signed-off-by: Georgi Djakov <quic_c_gdjako@quicinc.com>