clk: qcom: gcc-sun: Don't poll for PMU-controlled clocks to turn on
Some clocks are PMU-controlled, which can force the clocks off even if
SW is voting for them. The assumption being PMU can dynamically enable
the clocks as needed. In these cases, the clocks may not turn on when SW
actually votes for them, since PMU won't physically enable them until
later as-needed. So, skip polling for them to turn on in these cases.

Change-Id: I67ee76cab694fb50d0b651f48dc6c748b27006a5
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>