sched/walt: Modify trailblazer gating condition
When the trailblazer feature was introduced in commit
8ce94a28bf04 ("sched/walt: Boost ramp-up demands"), one of the two
conditions to consider a task for elevated demand was to ensure the
runtime of the task should be larger than or equal to the current
demand of the task. The commit text of the change mentioned the same.

However, the code was left with checking only if the runtime and the
current demand were equal. In addition to accounting for the case
where runtime is equal to demand, it is important to also account for
the condition when runtime is larger than demand because as the task is
ramping up, its average demand might be lower than the current runtime.
Without checking for runtime > demand, this condition would be ignored,
which could result in a delay in the task being boosted until the
demand catches up, despite the ramp up threshold being met.

Note that this change doesn't change anything if the window policy is
set such that demand = max(runtime, average), because that necessitates:

1024 >= demand >= runtime

In addition, it is important to note that core isolation doesn't look at
task demand for isolation/unisolation criteria. Therefore, even if there
is a trailblazer task present with inflated demand, but its raw runtime
is less than the busy threshold for isolation, the prime core(s) can get
isolated and cause the trailblazer task to run on non-prime CPUs.

Change-Id: Iaa50fb5bc144098310edfc1e6d1b0dd8eb94e62b
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>