soc: qcom: dcvs: Enable CPU freq stats
The platform protocol optionally provides statistics memory
region that is associated with the performance domain
management protocol. For a given performance domain, and for
each performance level in that domain, statistics in the
shared memory region track the number of times that the level
has been used and the amount of time that the domain has been
in that performance level. This patch adds support for providing
statistics to user space via debugfs.

Change-Id: I8cc00e9e4476e04fb38872012adae9a0cc95e5e2
Signed-off-by: Shivnandan Kumar <quic_kshivnan@quicinc.com>