power: supply: qti_battery_charger: Update ship mode entry sequence
Currently, charger FW configures and enables ship mode in parallel with
shutdown activity. This can cause a race condition leaving the device in
a bad state if the device is powering off while SW is still issuing SPMI
writes.

Update the ship mode entry sequence so that when the
"qcom,ship-mode-immediate" property is specified, a message will be sent
to the charger FW to configure the ship mode immediately after user sets
ship_mode_en. This avoids the race condition. It will have a physical
effect only after a subsequent shutdown is initiated by the
user(like before).

Change-Id: I6dc175da50d69acfdde5eef85cc6ed76aa2abfb9
Signed-off-by: Anjelique Melendez <quic_amelende@quicinc.com>