sched/walt: Improve pipeline tracing
Make two improvements on pipeline tracing.

1) In sched_pipeline_tasks, ensure the number of pipeline tasks are also
printed. This will help debug situations where a workload expects a
certain number of pipeline tasks to be in the system.

2) Ensure sched_pipeline_swapped is printed whenever a swap takes place.
Currently, this tracepoint only prints when a tasks between gold and
prime are swapped, but doesn't print when a gold task is promoted to
prime.

Change-Id: Ia20a7b91f0bac42b0f901d7273e8883d9f12e7e2
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>