soc: qcom: crm: Toggle CRMB_PT_TRIGGER properly
The CRMB_PT_TRIGGER register isn't VCD or "resource" specific. It's a
single, common register for the entire CRMB_PT. Thus, passing
resource_idx to write_crm_reg() result in an invalid address. The
logical resource_idx is always 0 for CRMB_PT_TRIGGER.

Additionally, CRMB_PT_TRIGGER is never currently cleared. Which means
subsequent requests wouldn't be triggered again since it needs a rising
edge to actually trigger the request. Thus, set and clear the trigger so
that it's zero again when we need to trigger something again.

Change-Id: I3b94fa42931a1974e7a6191b08af01d4f1290560
Signed-off-by: Mike Tipton <quic_mdtipton@quicinc.com>