rpmsg: glink: pkt: Decrement pages_to_map for leading/trailing
When mapping leading or trailing pages, total_bytes_to_map is updated,
and an additional page is added to pages_to_map. However, during the
cleanup of PTEs, this extra page is not removed. As a result, subsequent
mappings returns an EBUSY error since the previous PTE entry wasn’t
removed correctly.

To address this issue, modify the calculation of pages_to_map by
decrementing one page each in the case of valid leading or
trailing pages.

Change-Id: I1ec6aee0b82c4ef168a83f20641e0d6d5846fc27
Signed-off-by: Sarannya S <quic_sarannya@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>