sched/walt: Disable colocation placement boost on SOC
Disable the placement boosting aspect of coloction on latest target, as
all tasks already start on Golds. However, retain the frequency boosting
aspect.

grp->skip_min indicates whether a group is under colocation. skip_min
is then used to drive both the placement and the frequency boost
aspects of colocation. An SOC config is constructed to disable just
the placement (and load balancing) aspects of colocation.

Change-Id: Ie6242c0d840b2479ab1fcfca800d4071992cef68
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>