sched/walt: Introduce new adaptive boost hierarchy
Currently, there are two tunables associated with adaptive boost:
adaptive_low_freq and adaptive_high_freq. The feature operates in a
simple fashion: if a given frequency is less than the
adaptive_low_freq, then frequency is snapped up to
adaptive_low_freq. Otherwise, so long as the frequency is less than or
equal to the adaptive_high_freq, it is snapped up to the
adaptive_high_freq.

Introduce a new base level, namely, adaptive_level_1. In an effort
to maintain backwards compatibility, adaptive_low_freq and
adaptive_high_freq shall retain their names, but can be thought of as
equivalent to second and third levels respectively.

Change-Id: I0f2b686e6483a05138b003a0cce039f7c67d9b43
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>