sched/walt: Remove and replace nr_big_cpus usage
nr_big_cpus represents a count of all CPUs excluding smalls.
Previously, this was achieved by ignoring the first cluster, but that
assumption does not hold true across all SOCs.

Furthermore, the notion of nr_big_cpus is incorrect as well in this
context. Userspace specifies the cpus_for_pipeline, and the user should
therefore be limited to only adding as many pipeline tasks as specified
CPUs.

Therefore, remove nr_big_cpus, and instead directly use the weight of
the userspace specified mask to guard the max number of pipeline tasks.

Change-Id: Id3da4db981ea0d02a0489f6cfa8863bf97a2ef19
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>