power: supply: qti_battery_charger: Handle usb_icl_ua configurations
After USB enumeration, the USB driver votes for input current
limit (ICL) based on the adapter type (500/900 mA for SDP). This
value is cached (usb_icl_ua) so that when there is an ADSP SSR
or charger PDR, it can be used to vote again to the charger FW.

However, in battery_chg_update_usb_type_work(), we clear this
usb_icl_ua if the charger adapter type is not SDP or USB_PD.
With an SDP connected to DUT, when the charger PD restarts
after an SSR/PDR, it sends the first notification with unknown
USB adapter type followed by SDP. To handle this, ignore it when
the adapter type is unknown.

While at it, in battery_chg_update_usb_type_work(), add a check
for CDP adapter type before resetting usb_icl_ua to 0 as we
allow setting USB ICL for that adapter.

Also, in usb_psy_set_icl(), reset usb_icl_ua to 0 when it's
invoked for unknown adapter type to clear stale values if any.

Change-Id: Id8f1df2f3c3ca9992eed52487c996598c54552ff
Signed-off-by: Subbaraman Narayanamurthy <quic_subbaram@quicinc.com>