sched/walt: Introduce node for AUTO pipeline tasks
Add the "sched_pipeline_special" node, which is exposed to userspace.
Users can write the PID of a specific task to this node, ensuring that
the designated task will always be selected as part of the automatic
pipeline tasks. Regardless of dynamic task selection based on
utilization thresholds, this special userspace task remains included
in the AUTO pipeline use case.

Change-Id: I2ad58611bf6296d1b6dfbd08a925baacaaa9c85e
Signed-off-by: Shaleen Agrawal <quic_shalagra@quicinc.com>