gunyah: Fix possible overflow in the API gh_provide_user_mem
Assign the u32 to u16 which probably overflow when
it exceeds u16 limit. Fix it by change n_sgl_entries
type from u32 to unsigned long to avoid n_sgl_entries
is overflow when there are so many pages needed to do
hyp assgin. And check the n_sgl_entries is larger
than U16_MAX or not before assign to u16.

Change-Id: I49865fee5303d0437701b356014c67feb2d3c311
Signed-off-by: Peng Yang <quic_penyan@quicinc.com>