iio: adc: qcom-spmi-adc5-gen3: Reconfigure ADC_TM channel after fault
Currently, if an unexpected conversion fault occurs SW will only clear
the fault registers. If this conversion fault occurred for an
ADC TM channel, future TM measurements for that channel will fail
because the channel is no longer configured.

Update the interrupt handler so that when conversion faults are seen the
CONV_ERROR_STS register is read to determine which channel the fault
occurred on and how to properly reconfigure channel.

Change-Id: I9dd554594f773dcf025b38a1798851d6ce899d1c
Signed-off-by: Anjelique Melendez <quic_amelende@quicinc.com>