defconfig: sun: Disable CONFIG_QCOM_SPMI_ADC5_GEN3_DEBUG_LOGGING
CONFIG_QCOM_SPMI_ADC5_GEN3_DEBUG_LOGGING was enabled in sun
consolidate config so that register dumps can be collected if an ADC
conversion fault or timeout occurs along with triggering a panic
so that the register dump can be used for debugging.

However, because of some unlikely ADC TM conversion faults, this
is causing an issue. Disable this feature.

Signed-off-by: Anjelique Melendez <quic_amelende@quicinc.com>
Signed-off-by: Bruce Levy <quic_blevy@quicinc.com>