remoteproc: spss: handle sensors' regulator
Add support for sensor regulator configuration,
which is needed for certain targets like Sun.
Once SPSS is up, unvote for the regulator from HLOS
as SPSS votes for it.
Prior to SPSS SSR, vote to ensure SPSS can go up.
Once SPSS is up after SSR, unvote as SPSS votes
for the regulator.

Change-Id: Ic2dddcb2e6c146a6bc21a42545989f99bb13a34f
Signed-off-by: Nurit Lichtenstein <quic_nuritl@quicinc.com>