rpmsg: glink: Cleanup rx_done generation
The intents are now consumed in a sleepable context, and as such there
is no longer an issue with genrating rx_done messages from that same
context.

Remove the queue of done intents and the rx_done worker, and just handle
the processed intents directly in the qcom_glink_rx_done().

Change-Id: I469583faaa811c393e5c8982e7f3afdf57814c86
Signed-off-by: Bjorn Andersson <quic_bjorande@quicinc.com>
Signed-off-by: Tao Zhang <quic_taozhan@quicinc.com>