soc: qcom: rpmh: Optimize locking around flushing cache
In OS-Initiated mode of PSCI, a serialization is established in Linux
between the last CPU powering down the cluster / domain and the CPU
powering up back. The serialization is established by the genpd
framework and therefore it doesn't help to lock around the cache.

Refactor out the core of the flush functionality from the locking
wrappers. Calling the core functionality avoids a spin trylock, which
the RCU framework might consider suspicious under the circumstances.

Change-Id: I543672b1d00d77d4a6a899026e177f405d5652e3
Signed-off-by: Maulik Shah <quic_mkshah@quicinc.com>