power: supply: qti_battery_charger: Expose CHARGE_CONTROL_LIMIT{_MAX} properly
Currently, we are exposing the number of thermal levels for a board via
POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT_MAX property which is based on
the maximum allowed charging current and the number of thermal
mitigation levels or thermal mitigation step specified in the device
tree. Also, we are exposing the thermal mitigation level voted by
the thermal SW via POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT property.

As the thermal SW is going to set the current state for battery
thermal cooling device via set_cur_state() callback which is still
using the thermal levels, it would be good to expose the actual
charging current and the maximum allowed charging current that
comes from the charger firmware as-is via power supply properties.

Also, writing to POWER_SUPPLY_PROP_CHARGE_CONTROL_LIMIT property is
not needed anymore. Hence, updating the same.

Change-Id: Ia55abd342e1ecdd310da36a34afaffbea07dae85
Signed-off-by: Subbaraman Narayanamurthy <quic_subbaram@quicinc.com>