iommu-logger: Use device based lifecycle
Multiple devices may share the same iommu group. Save debug information
on a per-device basis when each device first attaches to a domain.

When arm-smmu is notified that a device is being removed, also remove
all saved debug information associated with that device.

This replaces the existing design, which only saved debug information
for the first device attached to a iommu domain.

Change-Id: I1ddcb7f14e9b0366b55063294207fc38e1b35943
Signed-off-by: Patrick Daly <quic_pdaly@quicinc.com>