ufs: qcom: Add support for partial CPUs device
In case a cpu is not populated or fused out, the cpu logical
to physical mapping will be rearranged. This may cause
unexpected behavior in the ufs QoS, CPU frequencies scaling,
and interrupt affinity logic. In such case, do not invoke
the QoS and UFS interrupt affinity features. The UFS performance
may be impacted for those devices. Check the sysfs node
sys/devices/system/cpu/possible for possible CPUs.

Change-Id: Ia6486064501c5b7c87b5a54293cfacdd34c554a9
Signed-off-by: Bao D. Nguyen <quic_nguyenb@quicinc.com>