hwmon: qti_amoled_ecm: Disable SDAM interrupts during ECM disable
When the display panel is turned off or goes blank, AMOLED ECM
would be disabled if the display-panels property is specified.
On PM8350B_ECM, this would be done with a peripheral register
write that disables ECM operation at the hardware. However on
PMD802X, ECM would be enabled always. When the display panel
goes off, this would cause ECM PBS to send SDAM interrupts
unnecessarily even when the samples are zeroes. Based on the
hardware recommendation, write 0 to ECM_WRITE_TO_SDAM so that
ECM PBS wouldn't send SDAM interrupts when ECM is disabled
by the SW.

Change-Id: Ib80612fbceb9009f81fbede8871e79071234e2da
Signed-off-by: Subbaraman Narayanamurthy <quic_subbaram@quicinc.com>