sched/walt: Fix error in SBT check
Currently, SBT checks for number of big tasks in the maximum possible
cluster. For some chipsets, the number of clusters are lesser than the
maximum possible cluster which makes the SBT check erroneous.
Fix this by ensuring SBT checks for number of big tasks in the last
cluster available for the chipset instead of the maximum cluster
possible.

Change-Id: I57b24178f5f89e2810c1c80f0f84512946026768
Signed-off-by: Sai Harshini Nimmala <quic_snimmala@quicinc.com>