remoteproc: qcom: Debug Enhancement
Kernel uses arch timer to get the counter and manage its time
while the subsystem firmware uses memtimer to get the counter
and manage its time. So, it is hard to align the time diff unless
there is a ramdump, ipc log, etc and sometimes only subsystem
ramdump elf and dmesg is available for debug. If cycle counter
can be added at the crash handler, it will be helpful to
sync memtime and ktime just with a kernel log.

Change-Id: I766a03b8dd993a6884d6fd792a5e3b7ed1e79e4d
Signed-off-by: Mukesh Ojha <quic_mojha@quicinc.com>