dmaengine: msm_gpi: Update gpi_terminate_all sequence
currently in gpi_terminate_all function, we are performing
stop, reset and restart the channels but we need to do reset
whenever the stop channel fails only.
So, changed the sequence in gpi_terminate_all to perform stop channel,
if any issues in stop then do reset and restart the channels.

Also, the start sequence is changed to first do rx start and then
tx start.

Signed-off-by: Chandana Kishori Chiluveru <quic_cchiluve@quicinc.com>
Change-Id: If5833d31f5d94d1eebc8f70cd0c1bcdc0210d9d6